chore: Reuse component
