+++ packages/applet/src/components/basic/Navbar.vue
