+++ packages/applet/src/modules/router/components/routes/Index.vue
@@ -4,7 +4,7 @@ import { Pane, Splitpanes } from 'splitpanes'
 import { getInspectorState, getInspectorTree, onInspectorStateUpdated, onInspectorTreeUpdated } from '@vue/devtools-core'
 import { parse } from '@vue/devtools-kit'
 import type { InspectorNodeTag, InspectorState } from '@vue/devtools-kit'
-import Navbar from '../Navbar.vue'
+import Navbar from '~/components/basic/Navbar.vue'
 import SelectiveList from '~/components/basic/SelectiveList.vue'
 import DevToolsHeader from '~/components/basic/DevToolsHeader.vue'
 import Empty from '~/components/basic/Empty.vue'
