+++ packages/applet/src/modules/router/components/timeline/Index.vue
@@ -1,109 +1,11 @@
 <script setup lang="ts">
-import { Pane, Splitpanes } from 'splitpanes'
-import { onAddTimelineEvent } from '@vue/devtools-core'
-import { computed, ref } from 'vue'
-
-import type { InspectorState, TimelineEvent } from '@vue/devtools-kit'
-import Navbar from '../Navbar.vue'
-import EventList from './EventList.vue'
-import Empty from '~/components/basic/Empty.vue'
-import RootStateViewer from '~/components/state/RootStateViewer.vue'
-import { createExpandedContext } from '~/composables/toggle-expanded'
-import DevToolsHeader from '~/components/basic/DevToolsHeader.vue'
-
-const { expanded: expandedStateNodes } = createExpandedContext('timeline-state')
-
-// event info + group info = [0, 1]
-expandedStateNodes.value = ['0', '1']
+import Timeline from '~/components/timeline/index.vue'
 
 const LAYER_ID = `router:navigations:0`
-const eventList = ref<TimelineEvent['event'][]>([])
-const groupList = ref<Map<number, TimelineEvent['event'][]>>(new Map())
-const selectedEventIndex = ref(0)
-const selectedEventInfo = computed(() => eventList.value[selectedEventIndex.value] ?? null)
-// event info
-const normalizedEventInfo = computed(() => {
-  const info: InspectorState[] = []
-  for (const key in selectedEventInfo.value?.data) {
-    info.push({
-      key,
-      type: key,
-      editable: false,
-      value: selectedEventInfo.value.data[key]!,
-    })
-  }
-  return info
-})
-// group info
-const normalizedGroupInfo = computed(() => {
-  const groupId = selectedEventInfo.value?.groupId
-  const groupInfo = groupList.value.get(groupId)!
-  if (groupInfo) {
-    const duration = groupInfo[groupInfo.length - 1]?.time - (groupInfo[0]?.time ?? 0)
-    return [{
-      key: 'events',
-      type: 'events',
-      editable: false,
-      value: groupInfo.length,
-    }, duration && {
-      key: 'duration',
-      type: 'duration',
-      editable: false,
-      value: `${duration}ms`,
-    }].filter(Boolean)
-  }
-  return undefined
-})
-
-// normalize display info
-const displayedInfo = computed(() => {
-  return { 'Event Info': normalizedEventInfo.value, ...(normalizedGroupInfo.value && { 'Group Info': normalizedGroupInfo.value }) } as unknown as Record<string, InspectorState[]>
-})
-
-function normalizeGroupList(event: TimelineEvent['event']) {
-  const groupId = event.groupId
-  if (groupId !== undefined) {
-    groupList.value.set(groupId, groupList.value.get(groupId) ?? [])
-    groupList.value.get(groupId)?.push(event)
-  }
-}
-
-onAddTimelineEvent((payload) => {
-  if (!payload)
-    return
-
-  const { layerId, event } = payload
-  if (layerId !== LAYER_ID)
-    return
-
-  eventList.value.push(event)
-  normalizeGroupList(event)
-})
+const DOC_LINK = 'https://router.vuejs.org/'
+const REPO_LINK = 'https://github.com/vuejs/router'
 </script>
 
 <template>
-  <div class="h-full flex flex-col">
-    <DevToolsHeader doc-link="https://router.vuejs.org/" github-repo-link="https://github.com/vuejs/router">
-      <Navbar />
-    </DevToolsHeader>
-    <template v-if="eventList.length">
-      <div class="flex-1 overflow-hidden">
-        <Splitpanes class="h-full">
-          <Pane border="r base" size="40" h-full>
-            <div h-full select-none overflow-scroll class="no-scrollbar">
-              <EventList v-model="selectedEventIndex" :data="eventList" />
-            </div>
-          </Pane>
-          <Pane size="60">
-            <div h-full select-none overflow-scroll class="no-scrollbar">
-              <RootStateViewer class="p3" :data="displayedInfo" node-id="" inspector-id="" :disable-edit="true" expanded-state-id="timeline-state" />
-            </div>
-          </Pane>
-        </Splitpanes>
-      </div>
-    </template>
-    <Empty v-else class="flex-1">
-      No events
-    </Empty>
-  </div>
+  <Timeline :layer-id="LAYER_ID" :doc-link="DOC_LINK" :repo-link="REPO_LINK" />
 </template>
